+++ src/main/java/com/hyfixes/HyFixes.java
@@ -2,6 +2,7 @@ package com.hyfixes;
 
 import com.hyfixes.commands.ChunkStatusCommand;
 import com.hyfixes.commands.ChunkUnloadCommand;
+import com.hyfixes.commands.FixCounterCommand;
 import com.hyfixes.commands.InteractionStatusCommand;
 import com.hyfixes.commands.WhoCommand;
 import com.hyfixes.listeners.CraftingManagerSanitizer;
@@ -189,9 +190,10 @@ public class HyFixes extends JavaPlugin {
     private void registerCommands() {
         getCommandRegistry().registerCommand(new ChunkStatusCommand(this));
         getCommandRegistry().registerCommand(new ChunkUnloadCommand(this));
+        getCommandRegistry().registerCommand(new FixCounterCommand(this));
         getCommandRegistry().registerCommand(new InteractionStatusCommand(this));
         getCommandRegistry().registerCommand(new WhoCommand());
-        getLogger().at(Level.INFO).log("[CMD] Registered /chunkstatus, /chunkunload, /interactionstatus, and /who commands");
+        getLogger().at(Level.INFO).log("[CMD] Registered /chunkstatus, /chunkunload, /fixcounter, /interactionstatus, and /who commands");
     }
 
     @Override